+++ lib/core/cubit/map/map_state.dart
@@ -33,8 +33,6 @@ class ShowedStopOnMap extends MapState {
 }
 
 class RouteUpdated extends MapState {
-
   @override
-  List<Object> get props => [
-      ];
+  List<Object> get props => [];
 }
