+++ lib/core/cubit/route/route_cubit.dart
@@ -12,9 +12,8 @@ import '../../entity/model/models.dart';
 import '../../exception/exceptions.dart';
 import '../../global/global_info.dart';
 import '../../repository/repositories.dart';
-import '../../selector/route_selectors.dart';
 import '../../service/services.dart';
-import '../../utils/route_utils.dart';
+import '../../utils/utils.dart';
 import '../cubits.dart';
 
 part 'route_state.dart';
