+++ lib/core/cubit/stop/stop_cubit.dart
@@ -11,8 +11,7 @@ import '../../entity/model/models.dart';
 import '../../exception/exceptions.dart';
 import '../../global/global_info.dart';
 import '../../repository/stop_repository.dart';
-import '../../selector/route_selectors.dart';
-import '../../utils/time_utils.dart';
+import '../../utils/general/time_utils.dart';
 import '../../utils/utils.dart';
 import '../cubits.dart';
