+++ lib/core/utils/general/platform_utils.dart
