+++ lib/core/utils/general/string_utils.dart
