+++ lib/page/map/widget/map_stop_list_tile.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+
 import '../../../core/entity/model/models.dart';
 import '../../stop_list/widget/stop_list_tile.dart';
 
@@ -11,12 +12,11 @@ class MapStopListTile extends StopListTile {
     bool isBeginNextStops = false,
   })  : assert(stop != null),
         super(
-          key: key,
-          stop: stop,
-          isUsingPro: true,
-          isNextStopSuggested: isNextStopSuggested,
-          isBeginNextStops: isBeginNextStops,
-          headerBackgroundColor: const Color(0xFF3AA348),
-          headerTextColor: Colors.white
-        );
+            key: key,
+            stop: stop,
+            isUsingPro: true,
+            isNextStopSuggested: isNextStopSuggested,
+            isBeginNextStops: isBeginNextStops,
+            headerBackgroundColor: const Color(0xFF3AA348),
+            headerTextColor: Colors.white);
 }
