+++ lib/page/stop_list/widget/pending_stops_tab_view.dart
@@ -11,19 +11,25 @@ class PendingStopsTabView extends StatelessWidget {
     Key key,
     @required this.stops,
     @required this.isUsingPro,
+    this.term,
   }) : super(key: key);
 
   final List<StopModel> stops;
   final bool isUsingPro;
+  final String term;
 
   @override
   Widget build(BuildContext context) {
+    final filteredStops = stops.where((stop) {
+      final description = stop?.location?.description ?? '';
+      return description.contains(term);
+    }).toList();
     return BlocBuilder<RouteCubit, RouteState>(
       builder: (_, state) => ListView.builder(
         itemBuilder: (_, i) => Column(
           children: [
             StopListTile(
-              stop: stops[i],
+              stop: filteredStops[i],
               isNextStopSuggested: i == 0,
               isUsingPro: isUsingPro,
               isBeginNextStops: i == 1,
@@ -34,7 +40,7 @@ class PendingStopsTabView extends StatelessWidget {
             ),
           ],
         ),
-        itemCount: stops.length,
+        itemCount: filteredStops.length,
       ),
     );
   }
