+++ lib/widget/core/global_cubits_widget_provider.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
 import 'package:flutter_bloc/flutter_bloc.dart';
-import '../../core/cubit/global_cubits.dart';
 
+import '../../core/cubit/global_cubits.dart';
 import 'connectivity_listener.dart';
 
 class GlobalCubitsWidgetProvider extends StatelessWidget {
