+++ lib/widget/stop/stop_location_info.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+
 import '../../core/entity/model/models.dart';
 import '../../core/utils/utils.dart';
