+++ test/core/utils/business/global_info_utils_test.dart
@@ -3,7 +3,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:gm_driver_lite/core/entity/enum/enums.dart';
 import 'package:gm_driver_lite/core/entity/model/undeliverable_code_model.dart';
 import 'package:gm_driver_lite/core/global/global_info.dart';
-import 'package:gm_driver_lite/core/selector/global_info_selectors.dart';
+import 'package:gm_driver_lite/core/utils/business/global_info_utils.dart';
 
 void main() {
   group('GlobalInfo selectors', () {
