+++ test/core/utils/business/stop_utils_test.dart
