+++ test/setup/setup_widgets.dart
@@ -1,7 +1,8 @@
-import 'package:gm_driver_lite/core/global/hive.dart';
 import 'package:hive/hive.dart';
-import 'package:path_provider/path_provider.dart';
 import 'package:intl/date_symbol_data_local.dart';
+import 'package:path_provider/path_provider.dart';
+
+import 'package:gm_driver_lite/core/global/hive.dart';
 
 void setupWidgetsForTesting() async {
   initializeDateFormatting('pt_BR', null);
